moved itemdefinition
